apt: Cleanup slot data when a process exits
When launching a system applet in Old 3DS mode, the Home Menu process
exits without finalizing on APT. Clear slot data accordingly on an
OnProcessExit function, called when a process exits.

This allows Old 3DS mode to exit a system applet back to the Home Menu
without hanging in the transition, due to APT falsely thinking that the
Home Menu is running.

This also removes the previous workaround for other system applets that
didn't finalize correctly, and it was being cleared once
StartSystemApplet was called.